Move builder to correct folder
